extract elastic
