+++ src/main/java/de/muenchen/dave/repositories/elasticsearch/CustomSuggestIndex.java
@@ -2,27 +2,19 @@ package de.muenchen.dave.repositories.elasticsearch;
 
 import de.muenchen.dave.domain.elasticsearch.CustomSuggest;
 import java.util.List;
-import org.springframework.stereotype.Service;
 
-@Service
-public class CustomSuggestIndex {
+public interface CustomSuggestIndex {
 
-    private final CustomSuggestIndexElasticRepository customSuggestIndexElasticRepository;
-
-    public CustomSuggestIndex(final CustomSuggestIndexElasticRepository customSuggestIndexElasticRepository) {
-        this.customSuggestIndexElasticRepository = customSuggestIndexElasticRepository;
-    }
-
-    public List<CustomSuggest> findAllByNameAndFkid(String name, String fkid) {
+    default List<CustomSuggest> findAllByNameAndFkid(String name, String fkid) {
         //TODO not implemented yet
         return null;
     }
 
-    public void deleteAllByFkid(String fkid) {
+    default void deleteAllByFkid(String fkid) {
         //TODO not implemented yet
     }
 
-    public void saveAll(List<CustomSuggest> suggestions) {
+    default void saveAll(List<CustomSuggest> suggestions) {
         //TODO not implemented yet
     }
