+++ src/main/java/de/muenchen/elasticimpl/CustomSuggestIndexElasticRepository.java
@@ -1,4 +1,4 @@
-package de.muenchen.dave.repositories.elasticsearch;
+package de.muenchen.elasticimpl;
 
 import de.muenchen.dave.domain.elasticsearch.CustomSuggest;
 import java.util.List;
